Improve coverage of `AndroidSpeechRecognizer`
