+++ mobile/src/androidTest/java/ch/epfl/sdp/mobile/test/infrastructure/speech/android/AndroidSpeechRecognizerFactoryTest.kt
@@ -2,7 +2,6 @@ package ch.epfl.sdp.mobile.test.infrastructure.speech.android
 
 import android.content.Context
 import android.speech.SpeechRecognizer
-import ch.epfl.sdp.mobile.infrastructure.speech.android.AndroidSpeechRecognizer
 import ch.epfl.sdp.mobile.infrastructure.speech.android.AndroidSpeechRecognizerFactory
 import io.mockk.every
 import io.mockk.mockk
@@ -24,38 +23,5 @@ class AndroidSpeechRecognizerFactoryTest {
     factory.createSpeechRecognizer()
 
     verify { SpeechRecognizer.createSpeechRecognizer(context) }
-  }
-
-  @Test
-  fun given_recognizer_when_destroy_then_destroysFramework() {
-    val framework = mockk<SpeechRecognizer>()
-    val recognizer = AndroidSpeechRecognizer(framework)
-    every { framework.destroy() } returns Unit
-
-    recognizer.destroy()
-
-    verify { framework.destroy() }
-  }
-
-  @Test
-  fun given_recognizer_when_startListening_then_startsFrameworkListening() {
-    val framework = mockk<SpeechRecognizer>()
-    val recognizer = AndroidSpeechRecognizer(framework)
-    every { framework.startListening(any()) } returns Unit
-
-    recognizer.startListening()
-
-    verify { framework.startListening(any()) }
-  }
-
-  @Test
-  fun given_recognizer_when_stopListening_then_stopsFrameworkListening() {
-    val framework = mockk<SpeechRecognizer>()
-    val recognizer = AndroidSpeechRecognizer(framework)
-    every { framework.stopListening() } returns Unit
-
-    recognizer.stopListening()
-
-    verify { framework.stopListening() }
   }
 }
